clean and separate Planner as much as possible for now from the old Project.
