+++ src/main/groovy/com/github/tng/vnv/planner/oldlcm/model/TestSuiteRequest.groovy
@@ -32,44 +32,19 @@
  * partner consortium (www.5gtango.eu).
  */
 
-package com.github.tng.vnv.planner.restmock
+package com.github.tng.vnv.planner.oldlcm.model
 
+import com.fasterxml.jackson.annotation.JsonProperty
+import groovy.transform.EqualsAndHashCode
+import io.swagger.annotations.ApiModelProperty
 
-import com.github.tng.vnv.planner.model.NsRequest
-import com.github.tng.vnv.planner.model.NsResponse
-import org.springframework.web.bind.annotation.*
+import javax.validation.constraints.NotNull
 
-@RestController
-class TestPlatformManagerMock {
+@EqualsAndHashCode
+class TestSuiteRequest {
 
-    Map<String, NsResponse> networkServiceInstances = [:]
-
-    void reset() {
-        networkServiceInstances.clear()
-    }
-
-    @PostMapping('/mock/tpm/requests')
-    NsResponse deployNsForTest(@RequestBody NsRequest nsRequest) {
-        def networkServiceInstance = new NsResponse(
-                instanceUuid: nsRequest.requestType == 'CREATE_SERVICE' ? UUID.randomUUID().toString() : nsRequest.instanceUuid,
-                serviceUuid: nsRequest.serviceUuid,
-                status: nsRequest.requestType == 'CREATE_SERVICE' ? 'CREATED' : 'TERMINATED',
-        )
-        networkServiceInstance.id=networkServiceInstance.instanceUuid
-        networkServiceInstances[networkServiceInstance.id] = networkServiceInstance
-        networkServiceInstance
-    }
-
-    @GetMapping('/mock/tpm/requests')
-    List<NsResponse> getDeployedNs() {
-        []
-    }
-
-    @GetMapping('/mock/tpm/requests/{requestId}')
-    NsResponse getNsForTest(@PathVariable('requestId') String requestId) {
-        def nsi=networkServiceInstances[requestId]
-        nsi.status = 'READY'
-        nsi
-    }
-
+    @ApiModelProperty(required = true)
+    @NotNull
+    @JsonProperty("test_uuid")
+    String testUuid
 }
