+++ src/main/groovy/com/github/tng/vnv/planner/oldlcm/model/TestSuiteResult.groovy
@@ -32,25 +32,35 @@
  * partner consortium (www.5gtango.eu).
  */
 
-package com.github.tng.vnv.planner.model
+package com.github.tng.vnv.planner.oldlcm.model
 
 import groovy.transform.EqualsAndHashCode
+import io.swagger.annotations.ApiModelProperty
+
+import javax.validation.constraints.NotNull
 
 @EqualsAndHashCode
-class TestPlanOld {
-    String uuid
+class TestSuiteResult {
     String packageId
-    List<NetworkServiceInstance> networkServiceInstances = []
-    List<TestSuiteResult> testSuiteResults = []
+    String uuid
+    String testPlanId
+    String instanceUuid
+    String serviceUuid
+
+    @ApiModelProperty(required = true)
+    @NotNull
+    String testUuid
+
     String status
+
 
     @Override
     public String toString() {
-        final StringBuffer sb = new StringBuffer("\n - TestPlanOld{ \n");
+        final StringBuffer sb = new StringBuffer("\n \tTestSuiteResult{");
         sb.append("uuid='").append(uuid).append('\'');
+        sb.append(", serviceUuid='").append(serviceUuid).append('\'');
+        sb.append(", testUuid='").append(testUuid).append('\'');
         sb.append(", status='").append(status).append('\'');
-        sb.append(", \nnsi's(#${networkServiceInstances.size()})=").append(networkServiceInstances);
-        sb.append(", \ntsr's(#${testSuiteResults.size()})=").append(testSuiteResults);
         sb.append('}');
         return sb.toString();
     }
