+++ src/main/groovy/com/github/tng/vnv/planner/repository/TestPlanRepository.groovy
@@ -1,0 +1,11 @@
+package com.github.tng.vnv.planner.repository
+
+
+import com.github.tng.vnv.planner.oldlcm.model.TestPlanOld
+
+interface TestPlanRepository {
+    def findNssByTestTag(String tag)
+    def findTssByTestTag(String tag)
+    def createTestPlan(TestPlanOld testPlan)
+    def updateTestPlan(TestPlanOld testPlan)
+}
