+++ src/main/groovy/com/github/tng/vnv/planner/repository/TestSuiteRepository.groovy
@@ -1,0 +1,11 @@
+package com.github.tng.vnv.planner.repository
+
+interface TestSuiteRepository {
+    def findByUuid(String uuid)
+
+    def save(def testSuite)
+
+    def update(def testSuite, String uuid)
+
+    def deleteByUuid(String l)
+}
