+++ src/main/groovy/com/github/tng/vnv/planner/service/TestSuiteService.groovy
@@ -1,16 +1,5 @@
 package com.github.tng.vnv.planner.service
 
-import groovy.util.logging.Log
-import org.springframework.stereotype.Component
-
-@Log
-@Component
-class TestSuiteService {
-    def findByUuid(String uuid) {}
-
-    def save(def testSuite) {}
-
-    def update(def testSuite, String uuid) {}
-
-    def deleteByUuid(String l) {}
+interface TestSuiteService {
+    def findByUuid(String uuid)
 }
