+++ src/main/groovy/com/github/tng/vnv/planner/service/TestSuiteServiceImpl.groovy
@@ -1,0 +1,13 @@
+package com.github.tng.vnv.planner.service
+
+import groovy.util.logging.Log
+import org.springframework.stereotype.Service
+
+@Log
+@Service("TestSuiteService")
+class TestSuiteServiceImpl implements TestSuiteService{
+
+
+    def findByUuid(String uuid) {}
+
+}
