+++ src/main/resources/application.yml
@@ -20,6 +20,7 @@ app:
       username: sonata
       password: 1234
     service.list.endpoint: ${app.gk.base.url}/services
+    service.list.by.tag.endpoint: ${app.gk.base.url}/services
     package.list.endpoint: ${app.gk.base.url}/packages
     service.metadata.endpoint: ${app.gk.base.url}/services/{networkServiceId}
     package.metadata.endpoint: ${app.gk.base.url}/packages/{packageId}
@@ -27,6 +28,7 @@ app:
     host: tng-api-gtw
     base.url: http://${app.vnvgk.host}:5000/api/v3
     test.list.endpoint: ${app.vnvgk.base.url}/tests/descriptors
+    test.list.by.tag.endpoint: ${app.vnvgk.base.url}/tests/descriptors
     test.metadata.endpoint: ${app.vnvgk.base.url}/tests/descriptors/{testUuid}
   cat:
     host: tng-cat
