+++ src/test/groovy/com/github/tng/vnv/planner/restmock/CuratorMock.groovy
@@ -35,41 +35,41 @@
 package com.github.tng.vnv.planner.restmock
 
 
-import com.github.tng.vnv.planner.model.TestPlanOld
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.PutMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RestController
+import com.github.tng.vnv.planner.model.NsRequest
+import com.github.tng.vnv.planner.model.NsResponse
+import org.springframework.web.bind.annotation.*
 
 @RestController
-class TestResultRepositoryMock {
+class CuratorMock {
 
-    Map<String, TestPlanOld> testPlans = [:]
-    def numOfCallsForTestPlanCreation = 0
-    def numOfCallsForTestPlanUpdate = 0
+    Map<String, NsResponse> networkServiceInstances = [:]
 
     void reset() {
-        testPlans.clear()
+        networkServiceInstances.clear()
     }
 
-    @GetMapping('/mock/trr/test-plans')
-    List listTestPlans() {
-        new ArrayList(testPlans.values())
+    @PostMapping('/mock/tpm/requests')
+    NsResponse deployNsForTest(@RequestBody NsRequest nsRequest) {
+        def networkServiceInstance = new NsResponse(
+                instanceUuid: nsRequest.requestType == 'CREATE_SERVICE' ? UUID.randomUUID().toString() : nsRequest.instanceUuid,
+                serviceUuid: nsRequest.serviceUuid,
+                status: nsRequest.requestType == 'CREATE_SERVICE' ? 'CREATED' : 'TERMINATED',
+        )
+        networkServiceInstance.id=networkServiceInstance.instanceUuid
+        networkServiceInstances[networkServiceInstance.id] = networkServiceInstance
+        networkServiceInstance
     }
 
-    @PostMapping('/mock/trr/test-plans')
-    TestPlanOld createTestPlan(@RequestBody TestPlanOld testPlan) {
-        ++numOfCallsForTestPlanCreation
-        testPlan.uuid = UUID.randomUUID().toString()
-        testPlans[testPlan.uuid] = testPlan
+    @GetMapping('/mock/tpm/requests')
+    List<NsResponse> getDeployedNs() {
+        []
     }
 
-    @PutMapping('/mock/trr/test-plans/{testPlanId:.+}')
-    TestPlanOld updatePlan(@RequestBody TestPlanOld testPlan, @PathVariable('testPlanId') String testPlanId) {
-        ++numOfCallsForTestPlanUpdate
-         testPlan.uuid = testPlanId
-        testPlans[testPlan.uuid] = testPlan
+    @GetMapping('/mock/tpm/requests/{requestId}')
+    NsResponse getNsForTest(@PathVariable('requestId') String requestId) {
+        def nsi=networkServiceInstances[requestId]
+        nsi.status = 'READY'
+        nsi
     }
+
 }
